requirements.txt
django
requests
django-sslserver
django-extensions
python-dotenv
yahoo_oauth
yahoo_fantasy_api